+++ ElGroupo.Web/Models/Messages/EventMessageModel.cs
@@ -9,7 +9,7 @@ namespace ElGroupo.Web.Models.Messages
     {
         public bool CanEdit { get; set; }
         public string PostedBy { get; set; }
-        public int PostedById { get; set; }
+        public long PostedById { get; set; }
         public DateTime PostedDate { get; set; }
         public string Subject { get; set; }
         public string MessageText { get; set; }
