+++ ElGroupo.Web/Models/Users/AutoCompleteModel.cs
@@ -7,9 +7,10 @@ namespace ElGroupo.Web.Models.Users
 {
     public class AutoCompleteModel
     {
-        public int Id { get; set; }
+        public long Id { get; set; }
         public string Name { get; set; }
         public string Email { get; set; }
+        
 
     }
 }
